Add CMake build file
This enables MSVC to open and build the project using vcpkg installed SDL2 (if available). Defaults to only building octo-cli otherwise.

The build file has been set up to mimic the existing Makefile as much as possible, so the C standard version is only specified for non-Apple Unix environments, and warning levels are set per platform.

MSVC does not use "treat warnings as errors" (/WX) due to a significant number of warnings regarding variable shadowing, functions with void parameters, and spectre vulnerabilities... Could probably go back to /W4 instead of /Wall but other platforms were using pedantic warnings so figured it was best to be consistent with the warning level more than how the compiler reacts.